Add missing default cases
Summary:
**Context:** https://fburl.com/switch-enum

----

Now that `-Wswitch-enum` is an error in all of `fbobjc`, making it such that all enum values need a case within a switch regardless of having a default case, we want to also enable `-Wswitch-default` as an error (ensuring we cover all values outside the given enum's values).  This will reduce SEVs as it will eliminate undefined behavior when a value outside the enum range is switched upon.

Differential Revision: D87663451

fbshipit-source-id: 3eee5dfe84462b9c8deaa35d79c2ecd2027ce8a5