Added preview app for IO::Serialization
Summary:
### Summary

This diff is related to the addition of a preview app for IO::Serialization in the Ocean project. The changes include:

*   Adding a new directory `xplat/ocean/res/application/ocean/demo/io/serialization/preview/osx/Assets.xcassets/` with a `Contents.json` file.
*   Creating a new file `PreviewMainWindow.h` in the `xplat/ocean/impl/application/ocean/demo/io/serialization/preview/win/` directory.
*   Adding a new file `Info.plist` in the `xplat/ocean/impl/application/ocean/demo/io/serialization/preview/osx/` directory.
*   Adding two new icon files (`ocean_icon_256x256.png` and `ocean_icon_128x128.png`) in the `xplat/ocean/res/application/ocean/demo/io/serialization/preview/osx/Assets.xcassets/AppIcon.appiconset/` directory.

These changes are likely part of a larger effort to create a preview application for demonstrating serialization functionality in the Ocean project. The addition of icon files and a `Contents.json` file suggests that this preview app will be deployed on macOS, while the `PreviewMainWindow.h` file indicates a Windows-specific implementation.

Reviewed By: enpe

Differential Revision: D88342203

fbshipit-source-id: 870cc6b61e7b80151e257c99aeb8bc0c31ebe88b