Added possibility to preserve coordinate frame properties during Bundle Adjustment
Summary:
**Change to Non-Linear Optimization Object Point**
==============================================

This diff includes several code changes to enhance the functionality of the non-linear optimization object point in the Ocean library.

### Addition of Absolute Transformation Header

The `AbsoluteTransformation.h` header file has been updated to include a minor correction in the documentation.

### Update to Non-Linear Optimization Object Point Implementation

In `NonLinearOptimizationObjectPoint.cpp`, an assertion has been added to ensure that the `subMatrixD` is not null, and a new header `AbsoluteTransformation.h` has been included.

### Addition of New Parameter to Optimize Object Points

The `NonLinearOptimizationObjectPoint.h` header file now includes a new parameter `applyAbsolutePoseAlignment` in the `optimizeObjectPoints` function to allow for absolute pose alignment during optimization.

Reviewed By: enpe

Differential Revision:
D89085241

Privacy Context Container: L1295962

fbshipit-source-id: 8b06c622e2ac1858be09c56fcc7449445c088e0d